fix: 경로 오류 및 constant 수정
