+++ src/pages/Admin/Class/index.tsx
@@ -2,30 +2,13 @@ import { useState } from 'react';
 
 import { tableColumnMap, tableColumns } from '@/constants/adminTableColumns';
 import useModal from '@/hooks/useModal';
+import DataItem from '@/pages/Admin/Class/components/DataItem';
 import { ClassTableData } from '@/types/admin';
 
 import ClassDetailModal from './components/ClassDetailModal';
 import ClassRegisterModal from './components/ClassRegisterModal';
 import SubClassRegisterModal from './components/SubClassRegisterModal';
 import useClass from './hooks/useClass';
-
-function DataItem({
-  column,
-  row,
-}: {
-  column: keyof ClassTableData;
-  row: ClassTableData;
-}) {
-  if (column === 'subClasses') {
-    const subClasses = row[column];
-    if (subClasses.length === 0) {
-      return <div>없음</div>;
-    }
-    const subClassesTitle = subClasses.map(subClass => subClass.title);
-    return <div className={'truncate'}>{subClassesTitle.join(', ')}</div>;
-  }
-  return <div className={'truncate'}>{row[column]}</div>;
-}
 
 function Index() {
   const [selectedClass, setSelectedClass] = useState<ClassTableData | null>(
